requirements.txt
aws-cdk.aws_cloudfront_origins
-e .